feat: [pss-cln-def-fold-validation-006] - Close delivery through required CI and PR merge
Rebase onto main and resolve mergeability blockers: merge wire boundary tests after composition fold, and exempt workers/taxonomy transitional shim from zero-statement functional coverage gates.
